put repo combinators into separate module
